requirements.txt
gradio==5.34.2
openai
python-dotenv
requests 